requirements.txt
requests==2.31.0
UnityPy==1.10.2
msgpack==1.0.5
pycryptodome==3.20.0
python-ffmpeg
pyinstaller

PyCriCodecs @ git+https://github.com/BlackNovaTech/PyCriCodecs@main

PyCriCodecs~=0.4.5
tqdm~=4.66.1
pillow~=10.2.0
